+++ domain/product.model.go
@@ -121,7 +121,7 @@ type ProductRequest1C struct {
 	MaterialCode        int      `gorm:"material_code" json:"material_code" validate:"required,gt=0"`
 	Name                string   `gorm:"name" json:"name" validate:"required,min=1,max=500"`
 	Manufacturer        string   `gorm:"manufacturer" json:"manufacturer" validate:"required,min=1,max=255"`
-	Quantity            int      `gorm:"quantity" json:"quantity" validate:"required,gte=0"`
+	Quantity            float64  `gorm:"quantity" json:"quantity" validate:"required,gte=0"`
 	RetailPrice         float64  `gorm:"retail_price" json:"retail_price"`
 	RetailPriceVat      float64  `gorm:"retail_price_vat" json:"retail_price_vat"`
 	SupplyPrice         float64  `gorm:"supply_price" json:"supply_price"`
@@ -129,7 +129,7 @@ type ProductRequest1C struct {
 	Sum                 float64  `gorm:"sum" json:"sum"`
 	VatPrice            float64  `gorm:"vat_price" json:"vat_price"`
 	Vat                 string   `gorm:"vat" json:"vat" validate:"required"`
-	Markup              int      `gorm:"markup" json:"markup"`
+	Markup              float64  `gorm:"markup" json:"markup"`
 	VatSum              float64  `gorm:"vat_sum" json:"vat_sum"`
 	ProductSeriesNumber string   `gorm:"product_series_number" json:"product_series_number" validate:"required"`
 	ExpireDate          string   `gorm:"expire_date" json:"expire_date" validate:"required"`
